requirements.txt
fastapi[standard]==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
alembic==1.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3